resolve the shortcuts, use WScript.Echo instead of MsgBox (so it can be run in command line using cscript)
1- Optionally resolve shortcuts to save a 
'									copy of the target file in the local 
'									backup folder. 
'								2- Changed all the msgboxes to be WScript.Echo
'									, this is more cscript friendly
'									, in case this is run in console mode. 
'								<TODO>: accept wildcards, like *.vbs...
'								<TODO>: if running in cscript, we can afford to be more verbose.
'								<TODO>: if running in cscript in console, we should be able 
'									to create switches (eg /ResolveShortcutFile, name of the backup folder...).